+++ packages/adp-tooling/src/writer/project-utils.ts
@@ -2,7 +2,7 @@ import { join } from 'path';
 import { readFileSync } from 'fs';
 import { v4 as uuidv4 } from 'uuid';
 import type { Editor } from 'mem-fs-editor';
-import type { CloudApp, AdpWriterConfig, CustomConfig, PackageJson } from '../types';
+import type { CloudApp, AdpWriterConfig, CustomConfig } from '../types';
 import {
     enhanceUI5DeployYaml,
     enhanceUI5Yaml,
@@ -12,15 +12,16 @@ import {
     enhanceUI5YamlWithTranspileMiddleware
 } from './options';
 
-import { UI5Config, getEsmTypesVersion, getTypesPackage } from '@sap-ux/ui5-config';
+import type { Package } from '@sap-ux/project-access';
 import type { OperationsType } from '@sap-ux/axios-extension';
+import { UI5Config, getEsmTypesVersion, getTypesPackage } from '@sap-ux/ui5-config';
 
 /**
  * Retrieves the package name and version from the package.json file located two levels up the directory tree.
  *
- * @returns {PackageJson} An object containing the `name` and `version` of the package.
+ * @returns {Package} An object containing the `name` and `version` of the package.
  */
-export function getPackageJSONInfo(): PackageJson {
+export function getPackageJSONInfo(): Package {
     const defaultPackage = {
         name: '@sap-ux/adp-tooling',
         version: 'NO_VERSION_FOUND'
@@ -42,13 +43,13 @@ export function getPackageJSONInfo(): PackageJson {
  * @param {string} pkg.version - The version of the tool generating the config.
  * @returns {CustomConfig} The generated ADP custom configuration object.
  */
-export function getCustomConfig(environment: OperationsType, { name: id, version }: PackageJson): CustomConfig {
+export function getCustomConfig(environment: OperationsType, { name: id, version }: Package): CustomConfig {
     return {
         adp: {
             environment,
             support: {
-                id,
-                version,
+                id: id ?? '',
+                version: version ?? '',
                 toolsId: uuidv4()
             }
         }