+++ packages/generator-adp/test/unit/questions/configuration.test.ts
@@ -1,12 +1,20 @@
 import type { ToolsLogger } from '@sap-ux/logger';
-import type { ListQuestion } from '@sap-ux/inquirer-common';
+import type { Manifest } from '@sap-ux/project-access';
 import type { AxiosError } from '@sap-ux/axios-extension';
+import type { ListQuestion } from '@sap-ux/inquirer-common';
 import { isAxiosError, type AbapServiceProvider } from '@sap-ux/axios-extension';
-import { FlexLayer, getConfiguredProvider, loadApps } from '@sap-ux/adp-tooling';
 import { getHostEnvironment, hostEnvironment } from '@sap-ux/fiori-generator-shared';
-import type { ConfigAnswers, TargetApplication, TargetSystems } from '@sap-ux/adp-tooling';
+import type { ConfigAnswers, SourceApplication, SystemLookup } from '@sap-ux/adp-tooling';
+import {
+    FlexLayer,
+    SourceManifest,
+    UI5VersionInfo,
+    getConfiguredProvider,
+    isAppSupported,
+    loadApps
+} from '@sap-ux/adp-tooling';
 
-import { initI18n } from '../../../src/utils/i18n';
+import { initI18n, t } from '../../../src/utils/i18n';
 import { configPromptNames } from '../../../src/app/types';
 import { ConfigPrompter } from '../../../src/app/questions/configuration';
 
@@ -23,7 +31,9 @@ jest.mock('@sap-ux/fiori-generator-shared', () => ({
 jest.mock('@sap-ux/adp-tooling', () => ({
     ...jest.requireActual('@sap-ux/adp-tooling'),
     getConfiguredProvider: jest.fn(),
-    loadApps: jest.fn()
+    loadApps: jest.fn(),
+    getSystemUI5Version: jest.fn().mockResolvedValue('1.135.0'),
+    isAppSupported: jest.fn()
 }));
 
 jest.mock('@sap-ux/axios-extension', () => ({
@@ -38,20 +48,18 @@ const logger: ToolsLogger = {
     debug: jest.fn()
 } as unknown as ToolsLogger;
 
-const getSystemInfoMock = jest.fn();
+const isAbapCloudMock = jest.fn();
 const provider = {
-    getLayeredRepository: jest.fn().mockReturnValue({
-        getSystemInfo: getSystemInfoMock
-    })
+    isAbapCloud: isAbapCloudMock
 } as unknown as AbapServiceProvider;
 
-const targetSystems: TargetSystems = {
+const sourceSystems: SystemLookup = {
     getSystems: jest.fn().mockResolvedValue([
         { Name: 'SystemB', Client: '200', Url: 'http://systemb.com', Authentication: 'Basic' },
         { Name: 'systemA', Client: '010', Url: 'http://systema.com', Authentication: 'NoAuthentication' }
     ]),
     getSystemRequiresAuth: jest.fn().mockResolvedValue(false)
-} as unknown as TargetSystems;
+} as unknown as SystemLookup;
 
 const dummyApps = [
     { id: 'app1', title: 'App One', ach: '', bspName: '', bspUrl: '', fileType: '', registrationIds: [] },
@@ -62,10 +70,11 @@ const dummyAnswers: ConfigAnswers = {
     system: 'SYS010',
     username: 'user1',
     password: 'pass1',
-    application: { id: 'app1', title: 'Some Title' } as unknown as TargetApplication
+    application: { id: 'app1', title: 'Some Title' } as unknown as SourceApplication
 };
 
 const loadAppsMock = loadApps as jest.Mock;
+const isAppSupportedMock = isAppSupported as jest.Mock;
 const isAxiosErrorMock = isAxiosError as unknown as jest.Mock;
 const getHostEnvironmentMock = getHostEnvironment as jest.Mock;
 const getConfiguredProviderMock = getConfiguredProvider as jest.Mock;
@@ -82,7 +91,7 @@ describe('ConfigPrompter Integration Tests', () => {
         getHostEnvironmentMock.mockReturnValue(hostEnvironment.vscode);
         loadAppsMock.mockResolvedValue(dummyApps);
         getConfiguredProviderMock.mockResolvedValue(provider);
-        configPrompter = new ConfigPrompter(targetSystems, layer, logger);
+        configPrompter = new ConfigPrompter(sourceSystems, layer, logger);
     });
 
     afterEach(() => {
@@ -160,7 +169,7 @@ describe('ConfigPrompter Integration Tests', () => {
                     statusText: 'Unauthorized'
                 }
             } as AxiosError;
-            getSystemInfoMock.mockRejectedValueOnce(axiosError);
+            isAbapCloudMock.mockRejectedValueOnce(axiosError);
             isAxiosErrorMock.mockReturnValueOnce(true);
 
             const prompts = configPrompter.getPrompts();
@@ -171,9 +180,36 @@ describe('ConfigPrompter Integration Tests', () => {
 
             expect(result).toEqual(`Authentication error: ${axiosError.message}`);
         });
+
+        it('system prompt validate should throw error when system info call fails with 405 and return true', async () => {
+            const axiosError = {
+                isAxiosError: true,
+                message: 'Method Not Allowed',
+                name: 'AxiosError',
+                response: {
+                    status: 405,
+                    statusText: 'Method Not Allowed'
+                }
+            } as AxiosError;
+            isAbapCloudMock.mockRejectedValueOnce(axiosError);
+            isAxiosErrorMock.mockReturnValueOnce(true);
+
+            const prompts = configPrompter.getPrompts();
+            const systemPrompt = prompts.find((p) => p.name === configPromptNames.system);
+            expect(systemPrompt).toBeDefined();
+
+            const result = await systemPrompt?.validate?.(dummyAnswers.system, dummyAnswers);
+
+            expect(result).toEqual(true);
+        });
     });
 
     describe('System CLI Validation Prompt', () => {
+        jest.spyOn(UI5VersionInfo, 'getInstance').mockReturnValue({
+            getSystemRelevantVersions: jest.fn(),
+            getRelevantVersions: jest.fn()
+        } as unknown as UI5VersionInfo);
+
         beforeEach(() => {
             getHostEnvironmentMock.mockReturnValue(hostEnvironment.cli);
         });
@@ -258,14 +294,62 @@ describe('ConfigPrompter Integration Tests', () => {
     });
 
     describe('Application Prompt', () => {
+        let getManifestSpy: jest.SpyInstance;
+
+        beforeEach(() => {
+            isAppSupportedMock.mockResolvedValue(true);
+            getManifestSpy = jest
+                .spyOn(SourceManifest.prototype, 'getManifest')
+                .mockResolvedValue({ 'sap.ui5': { flexEnabled: true } } as Manifest);
+            jest.spyOn(UI5VersionInfo, 'getInstance').mockReturnValue({
+                systemVersion: '1.135.0',
+                isVersionDetected: true
+            } as unknown as UI5VersionInfo);
+        });
+
         it('application prompt validate should return true if value is passed', async () => {
             const prompts = configPrompter.getPrompts();
             const appPrompt = prompts.find((p) => p.name === configPromptNames.application);
             expect(appPrompt).toBeDefined();
 
-            const result = appPrompt?.validate?.(dummyApps[0], dummyAnswers);
+            const result = await appPrompt?.validate?.(dummyApps[0], dummyAnswers);
 
             expect(result).toEqual(true);
+        });
+
+        it('application prompt validate should return string when manifest fetching fails', async () => {
+            const error = new Error('Test error');
+            getManifestSpy.mockRejectedValue(error);
+
+            const prompts = configPrompter.getPrompts();
+            const appPrompt = prompts.find((p) => p.name === configPromptNames.application);
+            expect(appPrompt).toBeDefined();
+
+            const result = await appPrompt?.validate?.(dummyApps[0], dummyAnswers);
+
+            expect(result).toEqual(error.message);
+        });
+
+        it('application prompt validate should return string when manifest is not found', async () => {
+            getManifestSpy.mockResolvedValue(undefined);
+            const prompts = configPrompter.getPrompts();
+            const appPrompt = prompts.find((p) => p.name === configPromptNames.application);
+            expect(appPrompt).toBeDefined();
+
+            const result = await appPrompt?.validate?.(dummyApps[0], dummyAnswers);
+
+            expect(result).toEqual(t('error.manifestCouldNotBeValidated'));
+        });
+
+        it('application prompt validate should return string when manifest flexEnabled is false', async () => {
+            getManifestSpy.mockResolvedValue({ 'sap.ui5': { flexEnabled: false } });
+            const prompts = configPrompter.getPrompts();
+            const appPrompt = prompts.find((p) => p.name === configPromptNames.application);
+            expect(appPrompt).toBeDefined();
+
+            const result = await appPrompt?.validate?.(dummyApps[0], dummyAnswers);
+
+            expect(result).toEqual(t('error.appDoesNotSupportAdaptation'));
         });
 
         it('application prompt validate should return string if value is not passed', async () => {
@@ -273,9 +357,8 @@ describe('ConfigPrompter Integration Tests', () => {
             const appPrompt = prompts.find((p) => p.name === configPromptNames.application);
             expect(appPrompt).toBeDefined();
 
-            const result = appPrompt?.validate?.(undefined, dummyAnswers);
-
-            expect(result).toEqual('Application has to be selected.');
+            const result = await appPrompt?.validate?.(undefined, dummyAnswers);
+            expect(result).toEqual(t('error.selectCannotBeEmptyError', { value: 'Application' }));
         });
     });
 
@@ -295,6 +378,39 @@ describe('ConfigPrompter Integration Tests', () => {
             const result = await (whenFn as (answers: ConfigAnswers) => Promise<boolean>)(dummyAnswers);
 
             expect(result).toEqual(false);
+        });
+
+        it('application validation cli prompt when should return false if application is undefined', async () => {
+            const prompts = configPrompter.getPrompts();
+            const appPrompt = prompts.find((p) => p.name === configPromptNames.appValidationCli);
+            expect(appPrompt).toBeDefined();
+
+            const whenFn = appPrompt?.when;
+            expect(typeof whenFn).toBe('function');
+
+            const result = await (whenFn as (answers: ConfigAnswers) => Promise<boolean>)({
+                ...dummyAnswers,
+                application: undefined as unknown as SourceApplication
+            });
+
+            expect(result).toEqual(false);
+        });
+
+        it('application validation cli prompt when should return false if manifest fetching fails', async () => {
+            isAppSupportedMock.mockResolvedValue(true);
+            const error = new Error('Test error');
+            jest.spyOn(SourceManifest.prototype, 'getManifest').mockRejectedValue(error);
+
+            const prompts = configPrompter.getPrompts();
+            const appPrompt = prompts.find((p) => p.name === configPromptNames.appValidationCli);
+            expect(appPrompt).toBeDefined();
+
+            const whenFn = appPrompt?.when;
+            expect(typeof whenFn).toBe('function');
+
+            await expect((whenFn as (answers: ConfigAnswers) => Promise<boolean>)(dummyAnswers)).rejects.toThrow(
+                error.message
+            );
         });
     });
 });